requirements.txt
Flask==3.1.1
pdfplumber==0.11.7
Werkzeug==3.1.3
gunicorn==21.2.0
python-docx==1.1.0